Initialize the filter with the "min-length" option specified
min_length:
Inspired by the "minLength" option in the Autocomplete Widget, the
"min-length" specifies  the minimum number of characters a user must
type before a search is performed.
"1" is the default value and is useful for local data with just a few
items, but a higher value should be used when a single character search
could match a few thousand items.